refactor: organizes Dialog exports in their respective folder
