+++ src/components/ui/dialog/index.ts
@@ -1,76 +1,3 @@
-import { render, createVNode } from 'vue';
-import DialogComponent from './Dialog.vue';
-import type { Variant } from '../../../types';
-
-export interface OpenDialogConfig {
-  id?: string;
-  title?: string;
-  hideFooter?: boolean;
-  message?: string;
-  onCallback?(val: string | boolean): void;
-  onClose?(val: string | boolean): void;
-  closeOnBackdrop?: boolean;
-  promptLabel?: string;
-  promptType?: string;
-  promptPlaceholder?: string;
-  cancelLabel?: string;
-  destructLabel?: string;
-  destructVariant?: Variant;
-  destructIcon?: string;
-  type?: 'prompt' | 'confirm';
-  opened?: boolean;
-  hideCancel?: boolean;
-}
-
-const globalConfig = {
-  opened: true,
-};
-
-let seed = 1;
-
-const open = (config: OpenDialogConfig = {}) => {
-  let dialogWrapper = null;
-  let dialogVM = null;
-  const id = 'ui-dialog-' + seed++;
-  dialogWrapper = document.createElement('div');
-  dialogWrapper.id = id;
-  dialogVM = createVNode(DialogComponent, {
-    ...globalConfig,
-    ...config,
-    id,
-  });
-
-  render(dialogVM, dialogWrapper);
-  document.body.appendChild(dialogWrapper);
-};
-const Dialog = {
-  open: open,
-  delete(config: OpenDialogConfig = {}) {
-    config.destructLabel = 'Deletar';
-    config.destructIcon = 'delete';
-    config.destructVariant = 'danger';
-    open(config);
-  },
-  prompt(config = {}) {
-    config = {
-      ...{
-        type: 'prompt',
-        destructVariant: 'danger',
-        destructLabel: 'Confirmar',
-      },
-      ...config,
-    };
-    open(config);
-  },
-  confirm(config: OpenDialogConfig = {}) {
-    config.type = 'confirm';
-    config.destructLabel = config.destructLabel || 'Confirmar';
-    config.destructIcon = 'check';
-    config.destructVariant = 'success';
-    config.hideCancel = true;
-    open(config);
-  },
-};
-
-export default Dialog;
-export { Dialog };
+export { default as Dialog } from './Dialog.vue';
+export { Dialog as $dialog } from './dialog';
+export type { DialogConfig, DialogProps } from './types';
